separated ui logic from the business logic
