+++ src/components/mod.rs
@@ -1,4 +1,2 @@
 mod business_components;
-pub mod ui_components; 
-
-
+pub mod ui_components;
